Renamed files

+++ R/init__cleanup_files.R
@@ -7,6 +7,10 @@
 #' @export
 
 cleanup_bici_files <- function(params) {
+    if (params$DEBUG) {
+        return()
+    }
+
     if (file.exists(params$config)) {
         message("Removed old BICI script")
         file.remove(params$config)
